test: removes redudant job to calculate git diff
