+++ .github/workflows/ci.yml
@@ -214,15 +214,6 @@ jobs:
 
   needs_e2e_build:
     uses: ./.github/workflows/needs-e2e-build.yml
-
-  debug_changed_files:
-    runs-on: ubuntu-latest
-    needs: needs_e2e_build
-    steps:
-      - name: Show changed files from needs_e2e_build
-        run: |
-          echo "Changed files:"
-          echo "${{ needs.needs_e2e_build.outputs.changed_files }}"
       
   build-android-apks:
     name: "Build Android APKs"
@@ -242,6 +233,8 @@ jobs:
       id-token: write
     needs: [needs_e2e_build, build-android-apks]
     uses: ./.github/workflows/run-e2e-smoke-tests-android.yml
+    with:
+      changed_files: ${{ needs.needs_e2e_build.outputs.changed_files }}
     secrets: inherit
 
   js-bundle-size-check:
